+++ publiccms-parent/publiccms-core/src/main/java/com/publiccms/controller/admin/sys/SysConfigDataAdminController.java
@@ -24,10 +24,17 @@ import com.publiccms.common.tools.RequestUtils;
 import com.publiccms.entities.log.LogOperate;
 import com.publiccms.entities.sys.SysConfigData;
 import com.publiccms.entities.sys.SysConfigDataId;
+import com.publiccms.entities.sys.SysDept;
+import com.publiccms.entities.sys.SysDeptConfigId;
 import com.publiccms.entities.sys.SysSite;
+import com.publiccms.entities.sys.SysUser;
 import com.publiccms.logic.component.config.ConfigComponent;
+import com.publiccms.logic.component.config.CorsConfigComponent;
+import com.publiccms.logic.component.site.EmailComponent;
 import com.publiccms.logic.service.log.LogLoginService;
 import com.publiccms.logic.service.sys.SysConfigDataService;
+import com.publiccms.logic.service.sys.SysDeptConfigService;
+import com.publiccms.logic.service.sys.SysDeptService;
 import com.publiccms.views.pojo.model.SysConfigParameters;
 
 /**
@@ -58,6 +65,17 @@ public class SysConfigDataAdminController extends AbstractController {
         }
         SysSite site = getSite(request);
         if (null != entity.getId()) {
+            SysUser user = ControllerUtils.getAdminFromSession(session);
+            SysDept dept = sysDeptService.getEntity(user.getDeptId());
+            if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                    || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                    || ControllerUtils
+                            .verifyCustom("noright",
+                                    !(dept.isOwnsAllConfig() || null != sysDeptConfigService
+                                            .getEntity(new SysDeptConfigId(user.getDeptId(), entity.getId().getCode()))),
+                                    model)) {
+                return CommonConstants.TEMPLATE_ERROR;
+            }
             entity.getId().setSiteId(site.getId());
             SysConfigData oldEntity = service.getEntity(entity.getId());
             if (null != oldEntity
@@ -70,19 +88,24 @@ public class SysConfigDataAdminController extends AbstractController {
             if (null != oldEntity) {
                 entity = service.update(oldEntity.getId(), entity, ignoreProperties);
                 if (null != entity) {
-                    logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
-                            LogLoginService.CHANNEL_WEB_MANAGER, "update.configData", RequestUtils.getIpAddress(request),
-                            CommonUtils.getDate(), JsonUtils.getString(entity)));
-                    configComponent.removeCache(site.getId(), entity.getId().getCode());
+                    logOperateService.save(
+                            new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER, "update.configData",
+                                    RequestUtils.getIpAddress(request), CommonUtils.getDate(), JsonUtils.getString(entity)));
                 }
             } else {
                 entity.getId().setSiteId(site.getId());
                 service.save(entity);
-                logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
-                        LogLoginService.CHANNEL_WEB_MANAGER, "save.configData", RequestUtils.getIpAddress(request),
-                        CommonUtils.getDate(), JsonUtils.getString(entity)));
-                configComponent.removeCache(site.getId(), entity.getId().getCode());
+                logOperateService
+                        .save(new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER, "save.configData",
+                                RequestUtils.getIpAddress(request), CommonUtils.getDate(), JsonUtils.getString(entity)));
             }
+            configComponent.removeCache(site.getId(), entity.getId().getCode());
+            if (emailComponent.getCode(site).equals(entity.getId().getCode())) {
+                emailComponent.clear(site.getId());
+            } else if (corsConfigComponent.getCode(site).equals(entity.getId().getCode())) {
+                corsConfigComponent.clear(site.getId());
+            }
+
         }
         return CommonConstants.TEMPLATE_DONE;
     }
@@ -92,7 +115,7 @@ public class SysConfigDataAdminController extends AbstractController {
      * @param _csrf
      * @param request
      * @param session
-     * @param model 
+     * @param model
      * @return view name
      */
     @RequestMapping("delete")
@@ -101,19 +124,37 @@ public class SysConfigDataAdminController extends AbstractController {
             return CommonConstants.TEMPLATE_ERROR;
         }
         SysSite site = getSite(request);
+        SysUser user = ControllerUtils.getAdminFromSession(session);
+        SysDept dept = sysDeptService.getEntity(user.getDeptId());
+        if (ControllerUtils.verifyNotEmpty("deptId", user.getDeptId(), model)
+                || ControllerUtils.verifyNotEmpty("deptId", dept, model)
+                || ControllerUtils
+                        .verifyCustom("noright",
+                                !(dept.isOwnsAllConfig()
+                                        || null != sysDeptConfigService.getEntity(new SysDeptConfigId(user.getDeptId(), code))),
+                                model)) {
+            return CommonConstants.TEMPLATE_ERROR;
+        }
         SysConfigData entity = service.getEntity(new SysConfigDataId(site.getId(), code));
         if (null != entity) {
             service.delete(entity.getId());
-            logOperateService.save(new LogOperate(site.getId(), ControllerUtils.getAdminFromSession(session).getId(),
-                    LogLoginService.CHANNEL_WEB_MANAGER, "delete.configData", RequestUtils.getIpAddress(request),
-                    CommonUtils.getDate(), JsonUtils.getString(entity)));
+            logOperateService.save(new LogOperate(site.getId(), user.getId(), LogLoginService.CHANNEL_WEB_MANAGER,
+                    "delete.configData", RequestUtils.getIpAddress(request), CommonUtils.getDate(), JsonUtils.getString(entity)));
             configComponent.removeCache(site.getId(), entity.getId().getCode());
         }
         return CommonConstants.TEMPLATE_DONE;
     }
 
     @Autowired
+    private SysDeptConfigService sysDeptConfigService;
+    @Autowired
+    private SysDeptService sysDeptService;
+    @Autowired
     private ConfigComponent configComponent;
+    @Autowired
+    private CorsConfigComponent corsConfigComponent;
+    @Autowired
+    private EmailComponent emailComponent;
     @Autowired
     private SysConfigDataService service;
 }
